pyproject.toml
[project]
name = "data-quality-dashboard"
version = "0.1.0"
description = "Data quality validation dashboard with semantic search"
requires-python = ">=3.12"
dependencies = [
    "polars>=1.0.0",
    "pandera[polars]>=0.21.0",
    "pydantic>=2.0.0",
    "fastapi[standard]>=0.100.0",
    "haystack-ai>=2.0.0",
    "uvicorn[standard]>=0.20.0",
    "pydantic-settings>=2.13.1",
    "python-json-logger>=4.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length    = 88
indent-width   = 4
target-version = "py312"
exclude        = [".git", ".venv", "__pycache__", "dist", "build"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "N",    # pep8-naming
    "PT",   # flake8-pytest-style
    "RUF",  # ruff-specific
]
ignore = [
    "E501",  # line-too-long (handled by formatter)
    "B008",  # function-call in default arg (FastAPI Depends pattern)
]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.ruff.format]
quote-style               = "double"
indent-style              = "space"
skip-magic-trailing-comma = false
line-ending               = "auto"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths    = ["tests"]

[dependency-groups]
dev = [
    "pytest-cov>=7.1.0",
]
